# docs/swagger.yaml
basePath: /api/v1
definitions:
  services_gateway_handler.AddFriendRequest:
    properties:
      friend_id:
        example: "123456"
        type: string
      reason:
        example: 想加你为好友
        type: string
    required:
    - friend_id
    type: object
...
  services_gateway_handler.CommonResponse:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  services_gateway_handler.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
    type: object
...
  services_gateway_handler.FriendRequestsResponse:
    properties:
      requests:
        items:
          $ref: '#/definitions/services_gateway_handler.FriendRequestInfo'
        type: array
    type: object
  services_gateway_handler.FriendsListResponse:
    properties:
      friends:
        items:
          $ref: '#/definitions/services_gateway_handler.FriendInfo'
        type: array
    type: object
...
  services_gateway_handler.ReplyFriendRequestBody:
    properties:
      accept:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: IM 即时通讯系统 Gateway 服务 API 文档，包含用户认证、好友管理、群组管理、消息系统等接口。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: IM 即时通讯系统 API
  version: "1.0"
paths:
  /friend_requests:
    get:
      consumes:
      - application/json
      description: 获取当前用户收到的所有待处理好友请求
      produces:
      - application/json
      responses:
        "200":
          description: 好友请求列表
          schema:
            $ref: '#/definitions/services_gateway_handler.FriendRequestsResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取待处理好友请求
      tags:
      - 好友
  /friend_requests/{request_id}:
    put:
      consumes:
      - application/json
      description: 接受或拒绝好友请求
      parameters:
      - description: 好友请求ID
        in: path
        name: request_id
        required: true
        type: string
      - description: 回复参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.ReplyFriendRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: 回复成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 回复好友请求
      tags:
      - 好友
  /friends:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有好友
      produces:
      - application/json
      responses:
        "200":
          description: 好友列表
          schema:
            $ref: '#/definitions/services_gateway_handler.FriendsListResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取好友列表
      tags:
      - 好友
    post:
      consumes:
      - application/json
      description: 发送好友请求给指定用户
      parameters:
      - description: 添加好友请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.AddFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 好友请求已发送
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 添加好友
      tags:
      - 好友
  /friends/{friend_id}:
    delete:
      consumes:
      - application/json
      description: 从好友列表中删除指定好友
      parameters:
      - description: 好友ID
        in: path
        name: friend_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除好友
      tags:
      - 好友
  /group/{id}/members/{member_id}:
    post:
      consumes:
      - application/json
      description: 群主修改群组成员的角色（管理员/普通成员）
      parameters:
      - description: 群组ID
        in: path
        name: id
        required: true
        type: string
      - description: 成员ID
        in: path
        name: member_id
        required: true
        type: string
      - description: 角色参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.ChangeGroupMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 修改群组成员角色
      tags:
      - 群组
  /group_requests:
    get:
      consumes:
      - application/json
      description: 群主获取待处理的群组加入请求列表
      parameters:
      - description: 群组ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.GetPendingGroupJoinRequestsBody'
      produces:
      - application/json
      responses:
        "200":
          description: 加入请求列表
          schema:
            $ref: '#/definitions/services_gateway_handler.GroupJoinRequestsResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取待处理群组加入请求
      tags:
      - 群组
    post:
      consumes:
      - application/json
      description: 发送加入群组的申请请求
      parameters:
      - description: 加入群组请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.JoinGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 申请已发送
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 申请加入群组
      tags:
      - 群组
  /group_requests/{request_id}:
    put:
      consumes:
      - application/json
      description: 群主接受或拒绝加入群组的请求
      parameters:
      - description: 请求ID
        in: path
        name: request_id
        required: true
        type: string
      - description: 回复参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.ReplyGroupJoinRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: 回复成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 回复群组加入请求
      tags:
      - 群组
  /groups:
    get:
      consumes:
      - application/json
      description: 获取当前用户加入的所有群组
      produces:
      - application/json
      responses:
        "200":
          description: 群组列表
          schema:
            $ref: '#/definitions/services_gateway_handler.UserGroupsResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户群组列表
      tags:
      - 群组
    post:
      consumes:
      - application/json
      description: 创建新的群组
      parameters:
      - description: 创建群组请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 群组创建成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CreateGroupResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建群组
      tags:
      - 群组
  /groups/{id}:
    get:
      consumes:
      - application/json
      description: 根据群组ID获取群组详细信息
      parameters:
      - description: 群组ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 群组信息
          schema:
            $ref: '#/definitions/services_gateway_handler.GroupInfo'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取群组信息
      tags:
      - 群组
  /groups/{id}/members:
    get:
      consumes:
      - application/json
      description: 获取指定群组的所有成员
      parameters:
      - description: 群组ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 群组成员列表
          schema:
            $ref: '#/definitions/services_gateway_handler.GroupMembersResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取群组成员列表
      tags:
      - 群组
  /groups/{id}/members/{member_id}:
    delete:
      consumes:
      - application/json
      description: 群主或管理员移除群组成员
      parameters:
      - description: 群组ID
        in: path
        name: id
        required: true
        type: string
      - description: 成员ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 移除群组成员
      tags:
      - 群组
  /groups/{id}/members/me:
    delete:
      consumes:
      - application/json
      description: 用户退出指定的群组
      parameters:
      - description: 群组ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/services_gateway_handler.CommonResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 退出群组
      tags:
      - 群组
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录获取 JWT Token
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回 Token
          schema:
            $ref: '#/definitions/services_gateway_handler.LoginResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      summary: 用户登录
      tags:
      - 用户认证
  /messages:
    post:
      consumes:
      - application/json
      description: 发送消息给指定用户或群组
      parameters:
      - description: 发送消息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services_gateway_handler.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 消息发送成功
          schema:
            $ref: '#/definitions/services_gateway_handler.SendMessageResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/services_gateway_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 发送消息
      tags:
      - 消息
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"